+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Added
+
+- Add option to console.warn when an non-existent step is parsed
+
 ## [1.0.0] 2025-12-17
 
 Initial release